+++ public/index.html
@@ -26,7 +26,7 @@
     -->
     <title>DaoSeeder</title>
   </head>
-  <body>
+  <body class="bg-white dark:bg-dark-background">
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
     <!--
